Move existing pre-purchase notices under the PrePurchaseNotices hierarchy
